requirements.txt
Flask>=3.0.0
flask-mysqldb>=1.1.0
Flask-Session>=0.5.0
Werkzeug>=3.0.0
